nonblock: Add Djukov MPMC with prepare-commit
This new queue replaces the old one as the basic MPMC queue.

It adds a lot of goodies, as:

-It's broken in prepare and commit blocks.
-Has a blocking fetch-add based enqueue that respects strict
 FIFO enqueuing order.
-Has a native mechanism for blocking the queue (Requires one
 bit from the counter).

This new queue is named "mpmc_b" and steals the name and
include file from the old one. The old value based one changes
name from "mpmc_b "to "mpmc_bt".

The old queue is exactly implemented in calls to this new one.

This is a breaking change. ssc will need to be updated.

The commit is very big too.