base: add array concat dynarray
Frequently dynamically sized one use arrays are needed but memory
allocations would gladly be avoided up to a certain array size that
can be assumed safe to be placed on the stack. This commit adds a
trivial wrapper to a fixed size array (to be on the stack most times)
concatenated with a dynamic array.