error: bl_file leaves errno of the syscall that failed
This is to be able to provide better diagnose messages on
file-related errors.

I am still disatisfied with the "bl_error" header, maybe I should
switch to POSIX errno and define it when not available or use a
struct containing both the bl_error and the errno value of what
failed.