error: BIG refactor: Don't lose errno's info
Until now bl was using its own error codes, which was relatively fine
but loses some error information available in the errno variable when
doing syscalls.

Switching to use the native errors on each platform isn't trivially
done on a multiplatform library, as there is very little standariza-
tion regarding the error codes on errno.

This commit converts "bl_err" into a struct containing the old bl
error code and _maybe_ the value of errno. This allows to have the
old error codes for error detection/flow control and errno for better
diagnose messages.

It also adds "bl_strerror".

Unfortunately this is a big breaking change and the semantics of
dealing with error codes becomes much uglier. I hope to not regret
it. It had to be done, so better sooner than later.