time_extras: Add cputstamp
This commit adds cputstamp, ftstamp and rewrites the conversion to
calendar clock logic.

"cputstamp" is an optimized counter that might or not be available
for a given platform. As of now it is implemented for x86-64 as
the rtdsc instruction. Functions to convert it to time and calendar
are added. The calendar function can be used for resyncing, as clocks
will unavoidably drift.

"ftstamp" is just a wrapper that is defined as "cputstamp" when
available, otherwise is just the regular "tstamp".

The conversion to calendar clock logic is rewritten. As it was found
that the previous version was an overkill. Selecting the samples that
repeat the most within is found to be enough empirically.