BREAK: Prefix everything with "bl_" or BL_
This BIG BREAKING commit refactors everything on the headers to be
prefixed with either "bl_" or "BL_".

Almost every function, datatype and macro is affected.

This is to feel safe when using this library on include headers for
other libraries of mine, e.g. malc. In such cases globally defining
macros like "pow2" is unnacceptable.

This was needed for a very long time, but posponed because it is very
tedious and unexciting work.