minimize failures of time based tests
These time-based smoke tests were spuriosly failing when run from
ninja, as it runs them on parallel.

This commit increases the times to test, so they become less
dependant on the system load.

It seems that the commit diff on time_test.c modifies every line
because the line endings changed from win to unix, which is good.